os compat

+++ graphics/test.cpp
@@ -342,7 +342,7 @@ private:
     delete pixels;
   }
 
-  void game_loop(const float game_time, const float frame_time)
+  void game_loop(const double game_time, const double frame_time)
   {
     //cout<<"dt: "<<frame_time<<endl;
     update_forces(game_time, frame_time);
